Moved opencl-compiler test from utilities to gocean1p0_opencl, this way the user sees a 'skipped test' notification if pylint was called with '--compile' instead of '--compileopencl'.
